test: Add a crate for test utilities
