+++ relay-server/Cargo.toml
@@ -75,4 +75,4 @@ libc = "0.2.71"
 
 [dev-dependencies]
 insta = "1.1.0"
-relay-log = { path = "../relay-log", features = ["test"] }
+relay-test = { path = "../relay-test" }
