+++ relay-server/src/utils/mod.rs
@@ -10,9 +10,6 @@ mod shutdown;
 mod timer;
 mod tracked_future;
 
-#[cfg(test)]
-mod test;
-
 #[cfg(feature = "processing")]
 mod kafka;
 #[cfg(feature = "processing")]
