+++ relay/src/main.rs
@@ -44,6 +44,7 @@
 //!  - [`relay-redis`]: Pooled Redis and Redis cluster abstraction.
 //!  - [`relay-sampling`]: Dynamic sampling functionality.
 //!  - [`relay-server`]: Endpoints and services.
+//!  - [`relay-test`]: Helpers for testing the web server and services.
 //!
 //! # Tools
 //!
@@ -70,6 +71,7 @@
 //! [`relay-redis`]: ../relay_redis/index.html
 //! [`relay-sampling`]: ../relay_sampling/index.html
 //! [`relay-server`]: ../relay_server/index.html
+//! [`relay-test`]: ../relay_test/index.html
 //! [`document-metrics`]: ../document_metrics/index.html
 //! [`generate-schema`]: ../generate_schema/index.html
 //! [`process-event`]: ../process_event/index.html
